+++ libs/libcmdtlm/cmd_tlm.hpp
@@ -34,7 +34,7 @@ public:
   CmdTlm(PacketReader *reader, PacketWriter *writer);
   virtual void telemetry(Commands &callback);
   virtual void control(const ControlPacketElement &e);
-  virtual void lwirFrame(const uint8_t frame[60][80]);
+  virtual void lwirFrame(const uint16_t frame[60][80]);
 };
 
 #endif
